shrink boot loader

+++ make.sh
@@ -1,4 +1,18 @@
 #!/bin/bash
+# ---------------------------------------------------------------------------
+# make - Make certikos kernel and build image
+
+# Hao Chen <hao.chen@yale.edu>
+
+# Usage: ./make.sh [rebuild] [gcc] [usb]
+#   rebuld: clean the compiled objects and make
+#   gcc: use gcc instead of compcertc to make
+#   usb: write the kernel to a usb stick
+
+# Revision history:
+# 2015-09-06  Created
+# ---------------------------------------------------------------------------
+
 
 yellow='\033[1;33m'
 green='\033[0;32m'
@@ -42,6 +56,7 @@ FLAGS="DEBUG_MSG=1 SERIAL_DEBUG=1 CONFIG_APP_VMM=1"
 
 need_clean="no"
 use_gcc="no"
+use_usb="no"
 
 for i in "$@"
 do
@@ -53,6 +68,9 @@ do
 		gcc)
 		use_gcc="yes"
 		shift
+		;;
+		usb)
+		use_usb="yes"
 		;;
 		*)
 		;;
@@ -92,42 +110,48 @@ else
 	num_ccomp=$( grep -c '+ ccomp' build.log )
 	num_ld=$( grep -c '+ ld' build.log )
 	num_as=$( grep -c '+ as' build.log )
+	byte_boot0=$(wc -c obj/boot/boot0 | awk '{print $1}')
+	byte_boot1=$(wc -c obj/boot/boot1 | awk '{print $1}')
+	byte_loader=$(wc -c obj/boot/loader | awk '{print $1}')
+	byte_kernel=$(wc -c obj/sys/kernel | awk '{print $1}')
 	title "[as: ${num_as}, cc: ${num_cc}, ccomp: ${num_ccomp}, ld: ${num_ld}]"
+	title "[boot0: ${byte_boot0}, boot1: ${byte_boot1}, loader: ${byte_loader}, kernel: ${byte_kernel}]"
 fi
 check $status
 
 title "build image"
-sudo ./build_image.py
+./make_image.py
 check $?
 
-title "check $disk is a usb stick"
-model=$( cat /sys/block/${disk}/device/model  )
+if [ "$use_usb" = "yes" ]; then
+	title "check $disk is a usb stick"
+	model=$( cat /sys/block/${disk}/device/model  )
+	
+	if [[ $model == "USB Flash Drive"* ]]; then
+		echo "yes"
+	else
+		echo "no, ${disk} is a ${model}. abort!"
+		exit 1
+	fi
+	title "write image to usb stick"
+	sudo dd if=certikos.img of=/dev/${disk} bs=2M conv=notrunc,noerror
+	check $?
 
-if [[ $model == "USB Flash Drive"* ]]; then
-	echo "yes"
+	eject_script="
+	#!/bin/bash
+	
+	sync
+	sudo sync
+	sudo eject ${disk}
+	"
+	echo -e "$eject_script" > eject.sh
+	chmod u+x eject.sh
+	
+	echo -e "please use ${green}./eject.sh${NC} to safely remove the USB stick"
 else
-	echo "no, ${disk} is a ${model}. abort!"
-	exit 1
+	title "convert to vmware hard disk"
+	qemu-img convert -p -O vmdk certikos.img certikos.vmdk
+	check $?
 fi
 
-title "convert to vmware hard disk"
-qemu-img convert -p -O vmdk certikos.img certikos.vmdk
-check $?
-
-title "write image to usb stick"
-sudo dd if=certikos.img of=/dev/${disk} bs=2M conv=notrunc,noerror
-check $?
-
 title "all done!"
-eject_script="
-#!/bin/bash
-
-sync
-sudo sync
-sudo eject ${disk}
-"
-echo -e "$eject_script" > eject.sh
-chmod u+x eject.sh
-
-echo -e "please use ${green}./eject.sh${NC} to safely remove the USB stick"
-
